Nullary function calls don't require parentheses.
Eg., "print" is now the same as "print()".

Before, giving the name of a function (that wasn't a self-call) didn't call
that function, it was just a reference to a function object, and parentheses
would turn it into a call on that object.  This was different from an object
call ("obj.func") or even a self-call.  It was the biggest "gotcha" when
writing programs in sqs.  And it wasn't particularly useful to be able to
access function objects.  So now a "bare" reference of a function is a nullary
call of that function in all cases, not just self-calls.