Modules: fix for cross-module superclasses.
Modules are now built in import order, instead of alphabetical order.  This
ensures that (where possible) a module is built before any of its classes are
used as a superclass of a class in another module.  When the child class is
built, that triggers building the superclass -- but we need to make sure the
superclass' module's block, with its locals-cum-module-variables, has been
built (or at least had its names resolved).